# app.py
from fastapi import FastAPI, HTTPException
from pydantic import BaseModel
from typing import Optional, List
from datetime import datetime
import logging

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')

app = FastAPI()

# Models
class ProvisionRequest(BaseModel):
    """Model for provisioning a new platform instance"""
    organization_name: str
    admin_email: str
    features: Optional[List[str]] = None
    region: str = "us-east-1"


# Routes
@app.get("/")
async def read_root():
    return {"Hello": "World"}


@app.get("/items/{item_id}")
async def read_item(item_id: int, q: Optional[str] = None):
    return {"item_id": item_id, "q": q}


@app.post("/pinksync/translate")
async def translate_text(text: str, target_language: str = "es"):
    """
    Translates the given text to the target language using PinkSync AI.
    """
    # Placeholder for actual translation logic
    translation = f"Translated '{text}' to {target_language} (PinkSync AI)"
    return {"translation": translation}


@app.post("/api/provision/unified-platform")
async def provision_unified_platform(request: ProvisionRequest):
    """
    Initiates the provisioning of a new unified communication platform instance.
    This endpoint would be called by mbtq.dev.
    """
    try:
        # In a real-world scenario, this would trigger a complex
        # asynchronous provisioning workflow.
        # This might involve:
        # 1. Creating a new database schema/instance for the organization.
        # 2. Deploying specific microservices or containers.
        # 3. Configuring AI models for the new instance.
        # 4. Setting up initial admin user accounts.
        # 5. Sending a confirmation email to the admin_email.

        instance_id = f"platform-{datetime.now().strftime('%Y%m%d%H%M%S')}-{request.organization_name.replace(' ', '-').lower()}"
        
        logging.info(f"Initiating provisioning for organization: {request.organization_name} (Instance ID: {instance_id})")
        logging.info(f"Admin Email: {request.admin_email}, Features: {request.features}, Region: {request.region}")

        # Simulate an asynchronous task
        # In a real system, you'd use a task queue (e.g., Celery, FastAPI BackgroundTasks)
        # to handle the actual provisioning process without blocking the API response.
        # For demonstration, we'll just log the action.

        return {
            "message": "Unified platform provisioning initiated. Check status for updates.",
            "instance_id": instance_id
        }
    except Exception as e:
        logging.error(f"Error during platform provisioning: {e}")
        raise HTTPException(status_code=500, detail=f"Failed to initiate provisioning: {str(e)}")